Update README.md
Finding and coordinating with others who share the same intersts can be difficult, especially when those interests can and should be experienced in a spontaneous spur-of-the-moment fashion. Imagine, for example, you feel like playing a pickup game of basketball, but you just didn't have the commitment required to sign up for an intermural sport. You can try to round up enough friends for a game and hope that everyones' schedules match up, but rarely does that approach work out. This could be said for any number of sports and activities alike. CoordinationStation matches up and groups users together based on their selected interests and schedules for a simple way to meet like-minded people and have fun without all of the arduous planning.